Add Kinect compiled hand tracker
Source code for this executable is missing due to lack of
professionalism by some people involved in the initial project.
The interface, though, it is very simple and it will be documented
further on.
Summarising: it is a basic Rock-Paper-Scissors hand gesture tracker
which sends the identified gesture ID for both hands over TCP to the
main program.